+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+# 5.0.1
+
+- feature: add some checks to reminders and sources that the Midi flag is actually true, mainly for some edge cases
+- feature: support conditional expressions when using Midi for sources (i.e. only show source if it evaluated true)
+- feature: rename Roll Dialog Colors setting menu to Button Style and update hint
+
 # 5.0.0
 
 - feature: support advantage roll mode features like overriding (i.e. use the Override change mode) and suppressing (i.e. use the Upgrade/Downgrade change mode)